third_party/blink/web_tests/virtual/webrtc-wpt-plan-b/external/wpt/webrtc/RTCPeerConnection-getStats.https-expected.txt

This is a testharness.js-based test.
PASS getStats() with no argument should succeed
PASS getStats(null) should succeed
PASS getStats() with track not added to connection should reject with InvalidAccessError
PASS getStats() with track added via addTrack should succeed
FAIL getStats() with track added via addTransceiver should succeed promise_test: Unhandled rejection with value: object "InvalidStateError: Failed to execute 'addTransceiver' on 'RTCPeerConnection': This operation is only supported in 'unified-plan'."
FAIL getStats() with track associated with both sender and receiver should reject with InvalidAccessError Failed to execute 'addTransceiver' on 'RTCPeerConnection': This operation is only supported in 'unified-plan'.
PASS getStats() with no argument should return stats report containing peer-connection stats on an empty PC
PASS getStats() track with stream returns peer-connection and outbound-rtp stats
PASS getStats() track without stream returns peer-connection and outbound-rtp stats
PASS getStats() audio outbound-rtp contains all mandatory stats
PASS getStats() video outbound-rtp contains all mandatory stats
FAIL getStats() audio and video validate all mandatory stats assert_equals: Expect dictionary.sdpFmtpLine to be string expected "string" but got "undefined"
PASS getStats() on track associated with RTCRtpSender should return stats report containing outbound-rtp stats
PASS getStats() on track associated with RTCRtpReceiver should return stats report containing inbound-rtp stats
PASS getStats() inbound-rtp contains all mandatory stats
FAIL getStats() with connected peer connections having tracks and data channel should return all mandatory to implement stats assert_unreached: test failed with error: Error: assert_equals: Expect dictionary.sdpFmtpLine to be string expected "string" but got "undefined" Reached unreachable code
FAIL getStats(track) should not work if multiple senders have the same track promise_test: Unhandled rejection with value: object "InvalidStateError: Failed to execute 'addTransceiver' on 'RTCPeerConnection': This operation is only supported in 'unified-plan'."
PASS RTCStats.timestamp increases with time passing
Harness: the test ran to completion.
